Add new binaries for VS2019
The new VisualStudioVersion number for VS2019 is 16.0. You may notice
that these new T4 binaries still have old version numbers on them (15.0,
11.0, 10.0). I tried changing the .targets file to just re-use the
existing 15.0 binaries in VS2019, but it did not work: there was a
reference mismatch between the older 15.0 binaries and these new ones.

As far as I know, these binaries /are/ the same as the old ones, but
with updated references and minor version numbers.

If some genius knows how to redirect the references instead of checking
in a whole new set of otherwise functionally identical binaries, I'd
love to have that, but for now I'm just going to do the easy thing that
works.